Deletion: Rename reaper2 to reaper; Closes #4462
