+++ bin/rucio-reaper
@@ -26,7 +26,7 @@
 import argparse
 import signal
 
-from rucio.daemons.reaper.reaper2 import run, stop
+from rucio.daemons.reaper.reaper import run, stop
 
 
 def get_parser():
