+++ lib/rucio/tests/test_abacus_collection_replica.py
@@ -39,7 +39,7 @@ from rucio.core.replica import delete_replicas, get_cleaned_updated_collection_r
 from rucio.core.rse import get_rse_id, add_rse
 from rucio.daemons.abacus import collection_replica
 from rucio.daemons.judge import cleaner
-from rucio.daemons.reaper import reaper2
+from rucio.daemons.reaper import reaper
 from rucio.daemons.undertaker import undertaker
 from rucio.db.sqla import models, session
 from rucio.db.sqla.constants import DIDType, ReplicaState
@@ -73,9 +73,9 @@ class TestAbacusCollectionReplica(unittest.TestCase):
         undertaker.run(once=True)
         cleaner.run(once=True)
         if cls.vo:
-            reaper2.run(once=True, include_rses='vo=%s&(%s)' % (cls.vo['vo'], cls.rse), greedy=True)
+            reaper.run(once=True, include_rses='vo=%s&(%s)' % (cls.vo['vo'], cls.rse), greedy=True)
         else:
-            reaper2.run(once=True, include_rses=cls.rse, greedy=True)
+            reaper.run(once=True, include_rses=cls.rse, greedy=True)
 
     def test_abacus_collection_replica_cleanup(self):
         """ ABACUS (COLLECTION REPLICA): Test if the cleanup procedure works correctly. """
@@ -143,13 +143,13 @@ class TestAbacusCollectionReplica(unittest.TestCase):
         assert str(dataset_replica['state']) == 'UNAVAILABLE'
 
         # Delete all files -> collection replica should be deleted
-        from rucio.daemons.reaper.reaper2 import REGION
+        from rucio.daemons.reaper.reaper import REGION
         REGION.invalidate()
         cleaner.run(once=True)
         if self.vo:
-            reaper2.run(once=True, include_rses='vo=%s&(%s)' % (self.vo['vo'], self.rse), greedy=True)
+            reaper.run(once=True, include_rses='vo=%s&(%s)' % (self.vo['vo'], self.rse), greedy=True)
         else:
-            reaper2.run(once=True, include_rses=self.rse, greedy=True)
+            reaper.run(once=True, include_rses=self.rse, greedy=True)
         self.rule_client.add_replication_rule([{'scope': self.scope, 'name': self.dataset}], 1, self.rse, lifetime=-1)
         collection_replica.run(once=True)
         dataset_replica = [replica for replica in self.replica_client.list_dataset_replicas(self.scope, self.dataset)]
