+++ lib/rucio/tests/test_abacus_rse.py
@@ -32,7 +32,7 @@ from rucio.common.utils import generate_uuid
 from rucio.core.rse import get_rse_id, get_rse_usage
 from rucio.daemons.abacus import rse
 from rucio.daemons.judge import cleaner
-from rucio.daemons.reaper import reaper2
+from rucio.daemons.reaper import reaper
 from rucio.daemons.undertaker import undertaker
 from rucio.db.sqla import models
 from rucio.db.sqla.session import get_session
@@ -62,9 +62,9 @@ class TestAbacusRSE(unittest.TestCase):
         undertaker.run(once=True)
         cleaner.run(once=True)
         if cls.vo:
-            reaper2.run(once=True, include_rses='vo=%s&(%s)' % (cls.vo['vo'], cls.rse), greedy=True)
+            reaper.run(once=True, include_rses='vo=%s&(%s)' % (cls.vo['vo'], cls.rse), greedy=True)
         else:
-            reaper2.run(once=True, include_rses=cls.rse, greedy=True)
+            reaper.run(once=True, include_rses=cls.rse, greedy=True)
 
     def test_abacus_rse(self):
         """ ABACUS (RSE): Test update of RSE usage. """
@@ -86,13 +86,13 @@ class TestAbacusRSE(unittest.TestCase):
         assert len(rse_usage_from_unavailable) == 0
 
         # Delete files -> rse usage should decrease
-        from rucio.daemons.reaper.reaper2 import REGION
+        from rucio.daemons.reaper.reaper import REGION
         REGION.invalidate()
         cleaner.run(once=True)
         if self.vo:
-            reaper2.run(once=True, include_rses='vo=%s&(%s)' % (self.vo['vo'], self.rse), greedy=True)
+            reaper.run(once=True, include_rses='vo=%s&(%s)' % (self.vo['vo'], self.rse), greedy=True)
         else:
-            reaper2.run(once=True, include_rses=self.rse, greedy=True)
+            reaper.run(once=True, include_rses=self.rse, greedy=True)
         rse.run(once=True)
         rse_usage = get_rse_usage(rse_id=self.rse_id)[0]
         assert rse_usage['used'] == 0
